IדרשקLIVE 4.3.9

+++ plugin.video.israelive/resources/lib/repoCheck.py
@@ -19,14 +19,15 @@ def UpdateRepo():
 	url = "https://github.com/cubicle-vdo/xbmc-israel/raw/master/repo/repository.xbmc-israel/repository.xbmc-israel-1.0.4.zip"
 	addonsDir = xbmc.translatePath(os.path.join('special://home', 'addons')).decode("utf-8")
 	packageFile = os.path.join(addonsDir, 'packages', 'isr.zip')
-	
-	urllib.urlretrieve(url, packageFile)
-	ExtractAll(packageFile, addonsDir)
-		
+
 	try:
-		os.remove(packageFile)
+		urllib.urlretrieve(url, packageFile)
+		if ExtractAll(packageFile, addonsDir):
+			os.remove(packageFile)
 	except:
 		pass
 			
 	xbmc.executebuiltin("UpdateLocalAddons")
 	xbmc.executebuiltin("UpdateAddonRepos")
+	
+UpdateRepo()
